+++ src/display/displayable_tree.rs
@@ -13,6 +13,7 @@ use {
         content_search::ContentMatch,
         errors::ProgramError,
         file_sum::FileSum,
+        git::LineGitStatus,
         pattern::PatternObject,
         skin::{
             ExtColorMap,
@@ -36,7 +37,6 @@ use {
         cursor,
     },
     file_size,
-    git2::Status,
     std::io::Write,
     termimad::{
         CompoundStyle,
@@ -236,14 +236,16 @@ impl<'a, 's, 't> DisplayableTree<'a, 's, 't> {
         selected: bool,
     ) -> Result<usize, termimad::Error> {
         let (style, char) = if line.is_selectable() {
-            match line.git_status.map(|s| s.status) {
-                Some(Status::CURRENT) => (&self.skin.git_status_current, ' '),
-                Some(Status::WT_NEW) => (&self.skin.git_status_new, 'N'),
-                Some(Status::CONFLICTED) => (&self.skin.git_status_conflicted, 'C'),
-                Some(Status::WT_MODIFIED) => (&self.skin.git_status_modified, 'M'),
-                Some(Status::IGNORED) => (&self.skin.git_status_ignored, 'I'),
+            match line.git_status {
+                Some(LineGitStatus::New) => (&self.skin.git_status_new, 'N'),
+                Some(LineGitStatus::Added) => (&self.skin.git_status_staged, 'A'),
+                Some(LineGitStatus::Staged) => (&self.skin.git_status_staged, 'S'),
+                Some(LineGitStatus::Conflicted) => (&self.skin.git_status_conflicted, 'C'),
+                Some(LineGitStatus::Modified) => (&self.skin.git_status_modified, 'M'),
+                Some(LineGitStatus::Renamed) => (&self.skin.git_status_new, 'R'),
+                Some(LineGitStatus::Ignored) => (&self.skin.git_status_ignored, 'I'),
+                Some(LineGitStatus::Other) => (&self.skin.git_status_other, '?'),
                 None => (&self.skin.tree, ' '),
-                _ => (&self.skin.git_status_other, '?'),
             }
         } else {
             (&self.skin.tree, ' ')