+++ src/tree_build/builder.rs
@@ -23,7 +23,6 @@ use {
         },
         tree::*,
     },
-    git2::Repository,
     id_arena::Arena,
     std::{
         collections::{
@@ -97,10 +96,7 @@ impl<'c> TreeBuilder<'c> {
         let line_status_computer = if options.filter_by_git_status || options.show_git_file_info {
             time!(
                 "init line_status_computer",
-                Repository::discover(&path)
-                    .ok()
-                    .as_ref()
-                    .and_then(LineStatusComputer::from),
+                LineStatusComputer::from_root(&path),
             )
         } else {
             None